New feature introduced for Game
