+++ game_frontend/src/index.js
@@ -12,13 +12,15 @@ const initialState = {
     movies: []
   },
   editor: {
-    code: '',
+    code: ''
+  },
+  gameView: {
     connectionParams: {
-        game_url_base: '',
-        game_url_path: '',
-        game_url_port: 0,
-        game_ssl_flag: false
-      }
+      game_url_base: '',
+      game_url_path: '',
+      game_url_port: 0,
+      game_ssl_flag: false
+    }
   }
 }
 const reduxStore = configureStore(initialState)
