+++ game_frontend/src/redux/features/Editor/actions.js
@@ -45,43 +45,11 @@ const keyPressed = code => (
   }
 )
 
-const getConnectionParamsRequest = gameID => (
-  {
-    type: types.GET_CONNECTION_PARAMS_REQUEST,
-    payload: {
-      gameID
-    }
-  }
-)
-
-const getConnectionParamsSuccess = connectionParams => (
-  {
-    type: types.GET_CONNECTION_PARAMS_SUCCESS,
-    payload: {
-      connectionParams
-    }
-  }
-)
-
-const emitUnityEvent = (gameObjectName, unityFunctionName, parameter) => (
-  {
-    type: types.EMIT_UNITY_EVENT,
-    payload: {
-      gameObjectName,
-      unityFunctionName,
-      parameter
-    }
-  }
-)
-
 export default {
   getCodeRequest,
   getCodeReceived,
   postCodeRequest,
   postCodeReceived,
   changeCode,
-  keyPressed,
-  getConnectionParamsRequest,
-  getConnectionParamsSuccess,
-  emitUnityEvent
+  keyPressed
 }
