+++ game_frontend/src/redux/features/Editor/epics.js
@@ -44,45 +44,8 @@ const changeCodeEpic = (action$, store, dependencies, scheduler = backgroundSche
     map(action => actions.changeCode(action.payload.code))
   )
 
-// TODO: maybe refactor all these catchErrors into individual actions instead of creating them here each time
-const getConnectionParamsEpic = (action$, store, { api }) => {
-  return action$.pipe(
-      ofType(types.GET_CONNECTION_PARAMS_REQUEST),
-      mergeMap(action => 
-        api.get(`games/${store.getState().game.id}/connection_params/`).pipe(
-          map(response => actions.getConnectionParamsSuccess(response)),
-          catchError(error => Observable.of({
-            type: types.GET_CONNECTION_PARAMS_FAIL,
-            payload: error.xhr.response,
-            error: true
-          }))
-        )
-      )
-    )
-  }
-
-const emitUnityEventEpic = (action$, store, { api }) => {
-  return action$.pipe(
-    tap(action => console.log(action)),
-    ofType(types.EMIT_UNITY_EVENT),
-    mergeMap(action =>
-      Observable.of(action).pipe(
-        api.emitUnityEvent,
-        map(event => ({ type: types.EMIT_UNITY_EVENT_SUCCESS })),
-        catchError(error => Observable.of({
-            type: types.EMIT_UNITY_EVENT_FAIL,
-            error: true
-          })
-        )
-      )
-    )
-  )
-}
-
 export default {
   getCodeEpic,
   postCodeEpic,
-  changeCodeEpic,
-  getConnectionParamsEpic,
-  emitUnityEventEpic
+  changeCodeEpic
 }
