+++ game_frontend/src/redux/features/Editor/types.js
@@ -9,14 +9,6 @@ const POST_CODE_FAILURE = 'features/Editor/POST_CODE_FAILURE'
 const CHANGE_CODE = 'features/Editor/CHANGE_CODE'
 const KEY_PRESSED = 'features/Editor/KEY_PRESSED'
 
-const GET_CONNECTION_PARAMS_REQUEST = 'features/Editor/GET_CONNECTION_PARAMS_REQUEST'
-const GET_CONNECTION_PARAMS_SUCCESS = 'features/Editor/GET_CONNECTION_PARAMS_SUCCESS'
-const GET_CONNECTION_PARAMS_FAIL = 'features/Editor/GET_CONNECTION_PARAMS_FAIL'
-
-const EMIT_UNITY_EVENT = 'features/Editor/EMIT_UNITY_EVENT'
-const EMIT_UNITY_EVENT_FAIL = 'features/Editor/EMIT_UNITY_EVENT_FAIL'
-const EMIT_UNITY_EVENT_SUCCESS = 'features/Editor/EMIT_UNITY_EVENT_SUCCESS'
-
 export default {
   GET_CODE_REQUEST,
   GET_CODE_SUCCESS,
@@ -25,11 +17,5 @@ export default {
   POST_CODE_SUCCESS,
   POST_CODE_FAILURE,
   CHANGE_CODE,
-  KEY_PRESSED,
-  GET_CONNECTION_PARAMS_REQUEST,
-  GET_CONNECTION_PARAMS_SUCCESS,
-  GET_CONNECTION_PARAMS_FAIL,
-  EMIT_UNITY_EVENT,
-  EMIT_UNITY_EVENT_FAIL,
-  EMIT_UNITY_EVENT_SUCCESS
+  KEY_PRESSED
 }
