+++ game_frontend/src/redux/features/Game/index.js
@@ -1,0 +1,7 @@
+import reducer from './reducers'
+
+export { default as gameEpics } from './epics'
+export { default as gameTypes } from './types'
+export { default as actions } from './actions'
+
+export default reducer
