+++ game_frontend/src/redux/features/Game/reducers.js
@@ -2,11 +2,10 @@ import types from './types'
 
 const editorReducer = (state = {}, action) => {
   switch (action.type) {
-    case types.GET_CODE_SUCCESS:
-    case types.CHANGE_CODE:
+    case types.GET_CONNECTION_PARAMS_SUCCESS:
       return {
         ...state,
-        code: action.payload.code
+        connectionParams: action.payload.connectionParams
       }
     default:
       return state
