+++ game_frontend/src/redux/features/Game/types.js
@@ -1,0 +1,16 @@
+const GET_CONNECTION_PARAMS_REQUEST = 'features/Editor/GET_CONNECTION_PARAMS_REQUEST'
+const GET_CONNECTION_PARAMS_SUCCESS = 'features/Editor/GET_CONNECTION_PARAMS_SUCCESS'
+const GET_CONNECTION_PARAMS_FAIL = 'features/Editor/GET_CONNECTION_PARAMS_FAIL'
+
+const EMIT_UNITY_EVENT = 'features/Editor/EMIT_UNITY_EVENT'
+const EMIT_UNITY_EVENT_FAIL = 'features/Editor/EMIT_UNITY_EVENT_FAIL'
+const EMIT_UNITY_EVENT_SUCCESS = 'features/Editor/EMIT_UNITY_EVENT_SUCCESS'
+
+export default {
+    GET_CONNECTION_PARAMS_REQUEST,
+    GET_CONNECTION_PARAMS_SUCCESS,
+    GET_CONNECTION_PARAMS_FAIL,
+    EMIT_UNITY_EVENT,
+    EMIT_UNITY_EVENT_FAIL,
+    EMIT_UNITY_EVENT_SUCCESS
+}
