+++ game_frontend/src/redux/features/index.js
@@ -2,15 +2,18 @@ import { combineReducers } from 'redux'
 import { combineEpics } from 'redux-observable'
 import ghibliReducer, { movieEpics } from './GhibliMovies'
 import editorReducer, { editorEpics } from './Editor'
+import gameReducer, { gameEpics } from './Game'
 
 const rootEpic = combineEpics(
   movieEpics,
-  ...Object.values(editorEpics)
+  ...Object.values(editorEpics),
+  ...Object.values(gameEpics)
 )
 
 const rootReducer = combineReducers({
   ghibli: ghibliReducer,
   editor: editorReducer,
+  gameView: gameReducer,
   game: (state = { id: 1 }) => state
 })
